refactor: roleplaying 모듈을 기능별 하위 디렉토리로 재구성
## 개요
app/roleplaying/ 디렉토리를 평면 구조에서 기능별 하위 디렉토리로 재구성했습니다.
모든 import 자동 업데이트 완료.

## 새로운 디렉토리 구조

app/roleplaying/
├── core/                      # 핵심 세션/상태 관리
│   ├── session_state_manager.py
│   ├── session_creation_builder.py
│   └── roleplaying_models.py
│
├── handlers/                  # WebSocket/메시지 처리
│   ├── ws_realtime_handler.py
│   ├── ws_message_models.py
│   ├── ws_message_router.py
│   └── session_validators.py
│
├── processing/                # 발화 처리
│   ├── user_utterance_processor.py
│   └── keyword_extraction_engine.py
│
├── api/                       # API 정의
│   ├── api_schemas.py
│   ├── roleplaying_router.py
│   └── roleplaying_repository.py
│
└── services/                  # 비즈니스 로직
    ├── llm/                   # LLM 서비스들 (10개 파일)
    ├── stt/                   # STT 서비스들 (4개 파일)
    ├── feedback/              # 피드백 평가 (2개 파일)
    ├── business/              # 시나리오/세션 생성 (4개 파일)
    ├── data/                  # 데이터 접근 (1개 파일)
    ├── utils/                 # 유틸리티 (3개 파일)
    └── dependencies/          # 의존성 주입 (4개 파일)

## 파일명 변경 (명확성 개선)
- llm_providers.py → llm_provider_factory.py
- interfaces.py → service_interfaces.py
- utils.py → service_utils.py
- title_utils.py → scenario_title_utils.py
- batch_stt_engine.py → batch_speech_to_text_engine.py
- streaming_stt_manager.py → streaming_speech_to_text_manager.py
- stt_service.py → speech_to_text_service.py
- repositories.py → data_repositories.py
- prompt_based_generator_service.py → prompt_scenario_generator_service.py
- utterance_processor.py → user_utterance_processor.py
- keyword_engine.py → keyword_extraction_engine.py
- message_router.py → ws_message_router.py
- session_manager.py → session_state_manager.py
- session_builder.py → session_creation_builder.py

## 이점
1. 명확한 기능별 그룹화 - 책임별로 서비스 분류
2. 자명한 구조 - 폴더명으로 목적 즉시 파악
3. 향상된 발견성 - 관련 코드가 함께 위치
4. 더 나은 네비게이션 - IDE가 관계를 더 잘 이해
5. 확장성 - 새 서비스 추가 용이
6. 팀 온보딩 - 신입이 구조를 즉시 이해

## 변경 사항
- 이동 파일: 44개
- 파일 이름 변경: 14개
- Import 업데이트: 60+개
- 새 디렉토리: 10개
- 새 __init__.py: 10개
