+++ app/roleplaying/core/roleplaying_models.py
