+++ app/roleplaying/core/session_creation_builder.py
