+++ app/roleplaying/core/session_state_manager.py
