+++ app/roleplaying/handlers/session_validators.py
@@ -81,8 +81,8 @@ from typing import Optional
 from fastapi import WebSocket
 
 from app.config import settings
-from app.roleplaying.session_manager import SessionStatus, SessionState, session_manager
-from app.roleplaying.ws_message_models import ErrorMessage
+from app.roleplaying.core.session_state_manager import SessionStatus, SessionState, session_manager
+from app.roleplaying.handlers.ws_message_models import ErrorMessage
 
 logger = logging.getLogger(__name__)
