+++ app/roleplaying/services/business/session_service_refactored.py
@@ -26,8 +26,8 @@ from typing import Optional
 from sqlalchemy.orm import Session
 
 from app.config import settings
-from app.roleplaying.api_schemas import ScenarioDetail
-from app.roleplaying.services.interfaces import SessionRepository, ScenarioRepository
+from app.roleplaying.api.api_schemas import ScenarioDetail
+from app.roleplaying.services.service_interfaces import SessionRepository, ScenarioRepository
 
 logger = logging.getLogger(__name__)
