+++ app/roleplaying/services/data/data_repositories.py
@@ -28,8 +28,8 @@ from sqlalchemy.orm import Session
 from sqlalchemy import text
 
 from app.config import settings
-from app.roleplaying.api_schemas import ScenarioDetail
-from app.roleplaying.services.interfaces import SessionRepository, ScenarioRepository
+from app.roleplaying.api.api_schemas import ScenarioDetail
+from app.roleplaying.services.service_interfaces import SessionRepository, ScenarioRepository
 
 logger = logging.getLogger(__name__)
