+++ app/roleplaying/services/llm/llm_conversation_analyzer.py
@@ -25,7 +25,7 @@ import logging
 from typing import Dict, Any, List
 
 from app.config import settings
-from app.roleplaying.services.llm_base import LLMServiceBase
+from app.roleplaying.services.llm.llm_base import LLMServiceBase
 from app.roleplaying.prompts.constants import CONVERSATION_ANALYSIS_PROMPT
 
 logger = logging.getLogger(__name__)
