+++ app/roleplaying/services/llm/llm_message_summarizer.py
@@ -25,7 +25,7 @@ import logging
 from typing import Dict, Any, List
 
 from app.config import settings
-from app.roleplaying.services.llm_base import LLMServiceBase
+from app.roleplaying.services.llm.llm_base import LLMServiceBase
 from app.roleplaying.prompts.constants import MESSAGE_SUMMARY_PROMPT
 
 logger = logging.getLogger(__name__)
