+++ app/roleplaying/services/llm/llm_scenario_enhancer.py
@@ -29,13 +29,13 @@ import logging
 from typing import Dict, Any, List
 
 from app.config import settings
-from app.roleplaying.services.llm_base import LLMServiceBase
+from app.roleplaying.services.llm.llm_base import LLMServiceBase
 from app.roleplaying.prompts.constants import (
     SITUATION_ENHANCEMENT_PROMPT,
     TITLE_GENERATION_PROMPT,
     PROMPT_QUESTIONS_PROMPT,
 )
-from app.roleplaying.services.utils import (
+from app.roleplaying.services.utils.service_utils import (
     extract_json_from_response,
     normalize_questions,
 )
