+++ app/roleplaying/services/service_interfaces.py
