+++ app/roleplaying/services/stt/audio_converter.py
