+++ app/roleplaying/services/stt/streaming_speech_to_text_manager.py
