+++ app/roleplaying/services/utils/azure_usage_tracker.py
