+++ app/roleplaying/services/utils/scenario_title_utils.py
