+++ app/roleplaying/services/utils/service_utils.py
